refactor: Reorganize package structure and file locations
