+++ app/src/androidTest/java/com/surivalcoding/composerecipeapp/presentation/component/BigButtonKtTest.kt
@@ -3,6 +3,7 @@ package com.surivalcoding.composerecipeapp.presentation.component
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
+import com.surivalcoding.composerecipeapp.ui.component.BigButton
 import org.junit.Assert.*
 import org.junit.Rule
 import org.junit.Test
