+++ app/src/main/java/com/surivalcoding/composerecipeapp/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp
+package com.surivalcoding.composerecipeapp.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
