+++ app/src/main/java/com/surivalcoding/composerecipeapp/ui/component/BigButton.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp.presentation.component
+package com.surivalcoding.composerecipeapp.ui.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
